% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{dispersion.plot}
\alias{dispersion.plot}
\title{Visualize a plot of the eigenvalue difference spectrum of a matrix or ensemble of matrices.}
\usage{
dispersion.plot(array, bins = 100)
}
\arguments{
\item{array}{a square matrix or matrix ensemble whose eigenvalues spacings are to be plotted}

\item{bins}{(optional) a string argument of the class of the matrix to label the plot title.}
}
\value{
A ggplot object containing a scatterplot of the matrix/matrix ensemble's eigenvalue spacings.
}
\description{
Returns a scatterplot of the eigenvalue spacings of a random matrix or ensemble.
}
\examples{
# Eigenvalue spacings plot of a normal matrix
P <- RM_norm(N = 5)
#dispersion.plot(P)

# Eigenvalue spacings plot of a beta matrix
Q <- RM_beta(N = 4, beta = 2)
#dispersion.plot(Q, mat_str = "Beta")

# Eigenvalue spacings plot of an ensemble of normal matrices
# ensemble <- RME_norm(N = 3, size = 10)
# dispersion.plot(ensemble)

}
