% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{dispersion}
\alias{dispersion}
\title{Obtain the dispersion of eigenvalues of a matrix or ensemble of matrices.}
\usage{
dispersion(array, components = T, norm = T)
}
\arguments{
\item{array}{a square matrix or matrix ensemble whose eigenvalues are to be returned}

\item{components}{returns the array with resolved real and imaginary components; otherwise returns complex-valued eigenvalues}

\item{norm}{use the norm metric for eigenvalue dispersion; otherwise returns absolute dispersion metric}
}
\value{
A tidy dataframe with the real & imaginary components of the eigenvalues and their norms along with a unique index.
}
\description{
Returns a vector of the eigenvalues dispersions of a random matrix or ensemble.
}
\examples{
# Eigenvalue spectrum of a matrix
P <- RM_norm(N = 5)
spectrum_P <- spectrum(P)

Q <- matrix(runif(2^2), ncol = 2)
spectrum_Q <- spectrum(Q)

# Eigenvalue spectra of ensemble matrices
ensemble <- RME_norm(N = 3, size = 10)
ensemble_spectrum <- spectrum(ensemble)

}
