% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R
\name{dispersion.scatterplot}
\alias{dispersion.scatterplot}
\title{Visualize a plot of the eigenvalue difference spectrum of a matrix or ensemble of matrices.}
\usage{
dispersion.scatterplot(array, metric = "id_diff_norm", ...)
}
\arguments{
\item{array}{a square matrix or matrix ensemble whose eigenvalue spacings are to be returned}

\item{metric}{a string denoting the eigenvalue dispersion metric (column of a dispersion object) to use}

\item{...}{any default-valued parameters taken as arguments by spectrum(array, ...)

Remove "no global binding for variable found" errors from using dplyr verbs by adding .data$ pronoun.}
}
\value{
A ggplot object containing a scatterplot of the matrix/matrix ensemble's eigenvalue spacings.
}
\description{
Returns a scatterplot of the eigenvalue spacings of a random matrix or ensemble.
}
\examples{
# Eigenvalue spacings plot of a normal matrix
P <- RM_norm(N = 5)
#dispersion.scatterplot(P)

# Eigenvalue spacings plot of a beta matrix
Q <- RM_beta(N = 4, beta = 2)
#dispersion.scatterplot(Q, mat_str = "Beta")

# Eigenvalue spacings plot of an ensemble of normal matrices
# ensemble <- RME_norm(N = 3, size = 10)
# dispersion.scatterplot(ensemble)

}
