% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R, R/parallel.R
\name{dispersion_parallel}
\alias{dispersion_parallel}
\title{Obtain the eigenvalue spacings of a matrix or ensemble of matrices.}
\usage{
dispersion_parallel(
  array,
  pairs = NA,
  sort_norms = TRUE,
  singular = FALSE,
  norm_pow = 1
)

dispersion_parallel(
  array,
  pairs = NA,
  sort_norms = TRUE,
  singular = FALSE,
  norm_pow = 1
)
}
\arguments{
\item{array}{a square matrix or matrix ensemble whose eigenvalue spacings are to be returned}

\item{pairs}{a string argument representing the pairing scheme to use}

\item{sort_norms}{sorts the eigenvalue spectrum by its norms when TRUE; otherwise, sorts eigenvalue by sign}

\item{norm_pow}{power to raise norm to - defaults to 1 (the standard absolute value); otherwise raises norm to the power of argument (beta norm)}
}
\value{
A tidy dataframe with the real & imaginary components of the eigenvalues and their norms along with a unique index.

A tidy dataframe with the real & imaginary components of the eigenvalues and their norms along with a unique index.
}
\description{
Returns a vector of the eigenvalue spacings of a random matrix or ensemble.

Returns a vector of the eigenvalue spacings of a random matrix or ensemble.
}
\examples{

# Eigenvalue dispersion in parallel
P <- RM_norm(N = 100, size = 500)
#disp_P <- dispersion_parallel(P)

# Eigenvalue dispersion in parallel
P <- RM_norm(N = 100, size = 500)
#disp_P <- dispersion_parallel(P)
}
