% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R, R/eigen.R
\name{.dispersion.varplot}
\alias{.dispersion.varplot}
\title{Visualize a plot of the variances of the eigenvalue dispersions by order difference level given a matrix or an ensemble.}
\usage{
.dispersion.varplot(array, ...)

.dispersion.varplot(array, ...)
}
\arguments{
\item{array}{a square matrix or matrix ensemble whose eigenvalue spacings are to be returned}

\item{...}{any default-valued parameters taken as arguments by spectrum(array, ...)}

\item{metric}{a string denoting the eigenvalue dispersion metric (column of a dispersion object) to use}
}
\value{
A ggplot object containing a scatterplot of the matrix/matrix ensemble's eigenvalue spacings.

A ggplot object containing a scatterplot of the matrix/matrix ensemble's eigenvalue spacings.
}
\description{
Returns a variance scatterplot of classes/levels of eigenvalue spacings of a random matrix or ensemble.

Returns a variance scatterplot of classes/levels of eigenvalue spacings of a random matrix or ensemble.
}
\examples{
# Eigenvalue spacings plot of a normal matrix
P <- RM_norm(N = 5)
#.dispersion.varplot(P)

# Eigenvalue spacings plot of a beta matrix
Q <- RM_beta(N = 4, beta = 2)
#.dispersion.varplot(Q, mat_str = "Beta")

# Eigenvalue spacings plot of an ensemble of normal matrices
# ensemble <- RME_norm(N = 3, size = 10)
# .dispersion.varplot(ensemble)

# Eigenvalue spacings plot of a normal matrix
P <- RM_norm(N = 5)
#.dispersion.varplot(P)

# Eigenvalue spacings plot of a beta matrix
Q <- RM_beta(N = 4, beta = 2)
#.dispersion.varplot(Q, mat_str = "Beta")

# Eigenvalue spacings plot of an ensemble of normal matrices
# ensemble <- RME_norm(N = 3, size = 10)
# .dispersion.varplot(ensemble)

}
