% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{RM_erdos}
\alias{RM_erdos}
\title{Generates a random stochastic matrix for a walk on an Erdos-Renyi graph
An Erdos-Renyi Graph is a graph whose edges are connected ~ Bern(p). Alternatively, this could be thought of as a stochastic matrix with parameterized sparsity.}
\usage{
RM_erdos(N, p, stoch = T)
}
\arguments{
\item{N}{number of dimensions of the square matrix}

\item{p}{the probability two vertices are connected in an Erdos-Renyi graph.}

\item{stoch}{indicates whether the matrix should be stochastic; changing this parameter may lead to the function returning invalid transition matrices.}
}
\value{
A random stochastic matrix corrosponding to a walk on an Erdos-Renyi graph with probability p.
}
\description{
Generates a random stochastic matrix for a walk on an Erdos-Renyi graph
An Erdos-Renyi Graph is a graph whose edges are connected ~ Bern(p). Alternatively, this could be thought of as a stochastic matrix with parameterized sparsity.
}
\examples{
P <- RM_stoch(N = 3)
P <- RM_stoch(N = 9, sparsity = TRUE)
Q <- RM_stoch(N = 9, symm = TRUE)
Q <- RM_stoch(N = 9, symm = TRUE, sparsity = TRUE)
}
