% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{RM_norm}
\alias{RM_norm}
\title{Generate a random matrix with normally distributed entries}
\usage{
RM_norm(N, mean = 0, sd = 1, symm = F, complex = F, hermitian = F)
}
\arguments{
\item{N}{Number of dimensions of the square matrix}

\item{mean}{Mean of entries}

\item{sd}{Standard deviation of entries}

\item{symm}{Indicates whether the matrix should be symmetric; equal to its transpose.}

\item{complex}{Indicates whether the matrix should have complex entries.}

\item{hermitian}{Indicates whether the matrix should be hermitian; equal to its conjugate transpose.}
}
\value{
A random matrix with normally distributed entries.
}
\description{
Generate a random matrix with normally distributed entries
}
\examples{
RM_norm(N = 3, mean = 1, sd = 2)
RM_norm(N = 3, symm = TRUE)
}
