% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{RM_stoch}
\alias{RM_stoch}
\title{Generate a random stochastic matrix}
\usage{
RM_stoch(N, symm = F, sparsity = F)
}
\arguments{
\item{N}{number of dimensions of the square matrix}

\item{symm}{indicates whether the matrix should be symmetric; equal to its transpose.}

\item{sparsity}{indicates whether the matrix should add some arbitrary sparsity (zeros)}
}
\value{
A random stochastic matrix.
}
\description{
Generate a random stochastic matrix
}
\examples{
P <- RM_stoch(N = 3)
P <- RM_stoch(N = 9, sparsity = TRUE)
Q <- RM_stoch(N = 9, symm = TRUE)
Q <- RM_stoch(N = 9, symm = TRUE, sparsity = TRUE)
}
