% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ensemble.R
\name{RME}
\alias{RME}
\title{Generate an ensemble of random matrices}
\usage{
RME(mat_type, args, ensemble_size)
}
\arguments{
\item{mat_type}{an RMAT random matrix function or identifier string}

\item{args}{a list of arguments to feed the RMAT random matrix function}

\item{ensemble_size}{the number of matrices to be simulated in the ensemble}
}
\value{
A list of matrices generated by the input matrix class and arguments.
}
\description{
Given a matrix class, this function returns an ensemble of that particular class of matrix.
  While random matrices usually do not exude unique properties on their own, they do indeed have
  deterministic properties (usually in spectral statistics) at the ensemble level.
}
\examples{
# Generate an ensemble of standard normal 3x3 matrices of size 20

## Using RMAT functions directly
ensemble <- RME(RM_norm, args = c(N = 3, mean = 0, sd = 1), ensemble_size = 20)

## Equivalent call
ensemble <- RME("norm", args = c(N = 3, mean = 0, sd = 1), ensemble_size = 20)

}
