% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{RME_beta}
\alias{RME_beta}
\title{Generate an ensemble of random beta matrices}
\usage{
RME_beta(N, beta, size)
}
\arguments{
\item{N}{number of dimensions of the square matrix}

\item{beta}{the value of the beta parameter for the beta ensemble}

\item{size}{the size of the ensemble (i.e. number of matrices)}
}
\value{
An ensemble (list) of beta matrices as specified by the matrix arguments.
}
\description{
Given the same arguments as RM_norm, this function returns an ensemble of that particular class of matrix.
  While random matrices usually do not exude unique properties on their own, they do indeed have
  deterministic properties at the ensemble level in terms of their spectral statistics.
}
\examples{

# Generate an ensemble of standard normal 3x3 matrices of size 20
# ensemble <-

}
