% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{spectrum.histogram}
\alias{spectrum.histogram}
\title{Visualize a plot of the eigenvalue distribution of a matrix or ensemble of matrices.}
\usage{
spectrum.histogram(array, ..., imaginary = F, bins = 100, mat_str = "")
}
\arguments{
\item{array}{a square matrix or matrix ensemble whose eigenvalues are to be plotted}

\item{...}{any default-valued parameters taken as arguments by spectrum(array, ...)}

\item{imaginary}{if TRUE, returns the distribution of imaginary components}

\item{bins}{number of bins of the histogram}

\item{mat_str}{(optional) a string argument of the class of the matrix to label the plot title.}
}
\value{
A ggplot object containing a histogram of the matrix/matrix ensemble's spectrum.
}
\description{
Returns a histogram of the eigenvalues of a random matrix or ensemble.
}
\examples{
# Eigenvalue spectrum plot of a matrix
P <- RM_norm(N = 5)
#spectrum.histogram(P)

# Labelled spectrum plot of a beta matrix
Q <- RM_beta(N = 4, beta = 2)
#spectrum.histogram(Q, mat_str = "Beta")

# Eigenvalue spectra plot of an ensemble of normal matrices
ensemble <- RME_norm(N = 3, size = 10)
#spectrum.histogram(ensemble)

}
