% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{spectrum.scatterplot}
\alias{spectrum.scatterplot}
\title{Visualize a plot of the eigenvalue spectrum of a matrix or ensemble of matrices.}
\usage{
spectrum.scatterplot(array, ..., mat_str = "")
}
\arguments{
\item{array}{a square matrix or matrix ensemble whose eigenvalues are to be returned}

\item{...}{any default-valued parameters taken as arguments by spectrum(array, ...)}

\item{mat_str}{(optional) a string argument of the class of the matrix to label the plot title.}
}
\value{
A ggplot object containing a scatterplot of the matrix/matrix ensemble's spectrum.
}
\description{
Returns a scatterplot of the eigenvalues of a random matrix or ensemble.
}
\examples{
# Eigenvalue spectrum plot of a matrix
P <- RM_norm(N = 5)
#spectrum.scatterplot(P)

# Labelled spectrum plot of a beta matrix
Q <- RM_beta(N = 4, beta = 2)
#spectrum.scatterplot(Q, mat_str = "Beta")

# Eigenvalue spectra plot of an ensemble of normal matrices
ensemble <- RME_norm(N = 3, size = 10)
#spectrum.scatterplot(ensemble)

}
