% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{spectrum_par}
\alias{spectrum_par}
\title{Obtain the eigenvalue spectrum of a matrix or ensemble of matrices concurrently (faster).}
\usage{
spectrum_par(
  array,
  components = TRUE,
  norm_order = TRUE,
  singular = FALSE,
  order = NA
)
}
\arguments{
\item{array}{a square matrix or matrix ensemble whose eigenvalues are to be returned}

\item{components}{returns the array with resolved real and imaginary components if TRUE; otherwise returns complex-valued eigenvalues}

\item{norm_order}{sorts the eigenvalue spectrum by its norms, otherwise sorts by sign}

\item{singular}{get the singular values of the matrix (i.e. square root of the eigenvalues of the matrix times its transpose)}

\item{order}{get eigenvalues with that given order (norm ranking); order 1 represents largest, order N represents smallest (where N is the number of eigenvalues).
If uninitialized, returns the entire spectrum.}
}
\value{
A tidy dataframe with the real & imaginary components of the eigenvalues and their norms along with a unique index.
}
\description{
Returns a tidied dataframe of the eigenvalues of a random matrix or ensemble concurrently (faster).
}
\examples{
# Generate a random matrix
P <- RME_norm(N = 20, size = 100)
# Compute the spectrum concurrently (faster)
#spec_P <- spectrum_par(P)

}
