% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{spectrum_plot}
\alias{spectrum_plot}
\title{Visualize a plot of the eigenvalue spectrum of a matrix or ensemble of matrices.}
\usage{
spectrum_plot(array, mat_str = "")
}
\arguments{
\item{array}{a square matrix or matrix ensemble whose eigenvalues are to be plotted}

\item{mat_str}{(optional) a string argument of the class of the matrix to label the plot title.}
}
\value{
A ggplot object containing a scatterplot of the matrix/matrix ensemble's spectrum.
}
\description{
Returns a scatterplot of the eigenvalues of a random matrix or ensemble.
}
\examples{
# Eigenvalue spectrum of a matrix
P <- RM_norm(N = 5)
spectrum_plot(P)

# Labelled spectrum plot of a beta matrix
Q <- RM_beta(N = 4, beta = 2)
spectrum_plot(Q, mat_str = "Beta")

# Eigenvalue spectra of ensemble matrices
ensemble <- RME("norm", args = c(N = 3), ensemble_size = 10)
spectrum_plot(ensemble)

}
