% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dispersion.R, R/eigen.R
\name{dispersion.histogram}
\alias{dispersion.histogram}
\title{Visualize a plot of the eigenvalue difference spectrum of a matrix or ensemble of matrices.}
\usage{
dispersion.histogram(array, ..., bins = 100)

dispersion.histogram(array, ..., bins = 100)
}
\arguments{
\item{array}{a square matrix or matrix ensemble whose eigenvalue spacings are to be returned}

\item{...}{any default-valued parameters taken as arguments by spectrum(array, ...)}

\item{bins}{(optional) a string argument of the class of the matrix to label the plot title.}

\item{metric}{a string denoting the eigenvalue dispersion metric (column of a dispersion object) to use}
}
\value{
A ggplot object containing a histogram of the matrix/matrix ensemble's eigenvalue spacings.

A ggplot object containing a histogram of the matrix/matrix ensemble's eigenvalue spacings.
}
\description{
Returns a histogram of the eigenvalue spacings of a random matrix or ensemble.

Returns a histogram of the eigenvalue spacings of a random matrix or ensemble.
}
\examples{
# Eigenvalue spacings plot of a normal matrix
P <- RM_norm(N = 5)
#dispersion.histogram(P)

# Eigenvalue spacings plot of a beta matrix
Q <- RM_beta(N = 4, beta = 2)
#dispersion.histogram(Q, mat_str = "Beta")

# Eigenvalue spacings plot of an ensemble of normal matrices
# ensemble <- RME_norm(N = 3, size = 10)
# dispersion.histogram(ensemble)

# Eigenvalue spacings plot of a normal matrix
P <- RM_norm(N = 5)
#dispersion.histogram(P)

# Eigenvalue spacings plot of a beta matrix
Q <- RM_beta(N = 4, beta = 2)
#dispersion.histogram(Q, mat_str = "Beta")

# Eigenvalue spacings plot of an ensemble of normal matrices
# ensemble <- RME_norm(N = 3, size = 10)
# dispersion.histogram(ensemble)

}
