% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{dispersion}
\alias{dispersion}
\title{Obtain the eigenvalue spacings of a matrix or ensemble of matrices.}
\usage{
dispersion(array, norm = T, diff_abs = F, components = T, digits = 3)
}
\arguments{
\item{array}{a square matrix or matrix ensemble whose eigenvalue spacings are to be returned}

\item{norm}{use the norm metric for eigenvalue spacing; otherwise returns absolute difference metric}

\item{diff_abs}{return the dispersions by computing the different of absolutes of the eigenvalues}

\item{components}{returns the array with resolved real and imaginary components; otherwise returns complex-valued vectors/distances}

\item{digits}{number of digits to round up values to}
}
\value{
A tidy dataframe with the real & imaginary components of the eigenvalues and their norms along with a unique index.
}
\description{
Returns a vector of the eigenvalue spacings of a random matrix or ensemble.
}
\examples{

# Eigenvalue dispersion of a normal matrix
P <- RM_norm(N = 5)
#disp_P <- dispersion(P)

# Eigenvalue dispersion of a stochastic matrix
Q <- RM_stoch(N = 5)
#disp_Q <- dispersion(Q)

# Eigenvalue dispersion of an ensemble
ensemble <- RME_norm(N = 3, size = 10)
#disp_ensemble <- dispersion(ensemble)

# Alternatively, use the pipe
#disp_ensemble <- RME_norm(N = 3, size = 10) \%>\% dispersion()

}
