% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{RM_beta}
\alias{RM_beta}
\title{Generate a Gaussian (Hermite) Beta Ensemble matrix with Dumitriu's Non-Invariant Tridiagonal Model}
\usage{
RM_beta(N, beta, cplx = F)
}
\arguments{
\item{N}{number of dimensions of the square matrix}

\item{beta}{the value of the beta parameter for the beta ensemble}

\item{cplx}{indicates whether the matrix should have complex entries.}
}
\value{
A random Hermite beta matrix with any integer parameter beta
}
\description{
Generate a Gaussian (Hermite) Beta Ensemble matrix with Dumitriu's Non-Invariant Tridiagonal Model
}
\examples{
P <- RM_beta(N = 3, beta = 4)
P <- RM_beta(N = 10, beta = 17)
}
