% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{RM_norm}
\alias{RM_norm}
\title{Generate normally distributed random matrix}
\usage{
RM_norm(N, mean = 0, sd = 1, symm = F, cplx = F, herm = F)
}
\arguments{
\item{N}{number of dimensions of the square matrix}

\item{mean}{mean of normal distribution of entries}

\item{sd}{standard deviation normal distribution of entries}

\item{symm}{indicates whether the matrix should be symmetric; equal to its transpose. Reserved for when complex = F, otherwise use hermitian = T.}

\item{cplx}{indicates whether the matrix should have complex entries.}

\item{herm}{indicates whether the matrix should be hermitian; equal to its conjugate transpose. Reserved for when complex = T, otherwise use hermitian = F.}
}
\value{
A random matrix with normally distributed entries.
}
\description{
Generate normally distributed random matrix
}
\examples{
P <- RM_norm(N = 3, mean = 1, sd = 2)
P <- RM_norm(N = 7, sd = 5, symm = TRUE)
Q <- RM_norm(N = 7, cplx = TRUE)
Q <- RM_norm(N = 5, mean = 0, cplx = TRUE, herm = TRUE)
}
