% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrices.R
\name{RME_erdos}
\alias{RME_erdos}
\title{Generate an ensemble of Erdos-Renyi transition matrices}
\usage{
RME_erdos(N, ..., size)
}
\arguments{
\item{N}{number of dimensions of the square matrix}

\item{...}{any default-valued parameters taken as arguments by the RM_ variant of this function}

\item{size}{the number of matrices to be simulated in the ensemble}
}
\value{
An ensemble (list) of Erdos-Renyi transition matrices as specified by the matrix arguments.
}
\description{
Given the same arguments as RM_norm, this function returns an ensemble of that particular class of matrix.
  While random matrices usually do not exude unique properties on their own, they do indeed have
  deterministic properties (usually in spectral statistics) at the ensemble level.
}
\examples{

# Generate an ensemble of standard normal 3x3 matrices of size 20
# ensemble <-

}
