% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.R
\name{spectrum}
\alias{spectrum}
\title{Obtain the eigenvalue spectrum of a matrix or ensemble of matrices.}
\usage{
spectrum(array)
}
\arguments{
\item{array}{a square matrix or matrix ensemble whose eigenvalues are to be returned}
}
\value{
A tidy dataframe with the real & imaginary components of the eigenvalues and their norms along with a unique index.
}
\description{
Returns a tidied dataframe of the eigenvalues of a random matrix or ensemble.
}
\examples{
# Eigenvalue spectrum of a matrix
P <- RM_norm(N = 5)
spectrum_P <- spectrum(P)

Q <- matrix(runif(2^2), ncol = 2)
spectrum_Q <- spectrum(Q)

# Eigenvalue spectra of ensemble matrices
#ensemble <- RME("norm", args = c(N = 3), ensemble_size = 10)
#ensemble_spectrum <- spectrum(ensemble)

}
