% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectrum.R
\name{spectrum}
\alias{spectrum}
\title{Obtain the eigenvalue spectrum of a matrix or ensemble of matrices.}
\usage{
spectrum(array, components = TRUE, sortByNorm = NA, order = NA)
}
\arguments{
\item{array}{a square matrix or matrix ensemble whose eigenvalues are to be returned}

\item{components}{returns the array with resolved real and imaginary components if TRUE; otherwise returns complex-valued eigenvalues}

\item{sortByNorm}{sorts the eigenvalue spectrum by its norms, otherwise sorts by sign}

\item{order}{get eigenvalues with that given order (norm ranking); order 1 represents largest, order N represents smallest (where N is the number of eigenvalues).
If uninitialized, returns the entire spectrum.}
}
\value{
A tidy dataframe with the real & imaginary components of the eigenvalues and their norms along with a unique index.
}
\description{
Returns a tidied dataframe of the eigenvalues of a random matrix or ensemble.
}
\examples{

# Eigenvalue spectrum of a random normal matrix
P <- RM_norm(N = 5)
spectrum_P <- spectrum(P)

Q <- matrix(runif(2^2), ncol = 2)
spectrum_Q <- spectrum(Q)

# Eigenvalue spectra of ensemble matrices
ensemble <- RME_norm(N = 3, size = 10)
ensemble_spectrum <- spectrum(ensemble)

}
