\name{summary,tilingSet-method}
\docType{methods}
\alias{summary,tilingSet-method}
\title{Summary Method for MAT Object}

\description{
This methods summarize tilinSet object}

\usage{
\S4method{summary}{tilingSet}(object)
}

\arguments{
\item{object}{A \code{tilingSet} object.}
}

\details{
This function will give a basic summary of a tilingSet object including chromosome/genome information.
}


\author{
  Charles Cheung, \email{cykc@interchange.ubc.ca} and
  Raphael Gottardo, \email{raphael.gottardo@ircm.qc.ca}
  Arnaud Droit, \email{arnaud.droit@ircm.qc.ca}
}

\seealso{
\code{\link{BPMAPCelParser}}, \code{\link{NormalizeProbes}}
}

\keyword{print}
